增加验证码

+++ myself-web/src/main/java/com/sipingsoft/core/shiro/ShiroRelamService.java
@@ -2,6 +2,11 @@ package com.sipingsoft.core.shiro;
 
 import java.util.Set;
 
+/**
+ * 权限获取接口
+ * @author He Chunxiao
+ * @since  2018-11-1 9:36
+ */
 public interface ShiroRelamService {
 	
 	/**
@@ -9,5 +14,5 @@ public interface ShiroRelamService {
      * @param userId
      * @return
      */
-    Set<String> findPermissions(Long userId);
+    Set<String> findPermissions(Integer userId);
 }
